requirements.txt
torch
transformers
datasets
trl
accelerate
pyyaml